refactor : InputValidator를 각각의 use case로 분류하고 클래스 분리
