+++ src/main/kotlin/domain/validator/ValidateUseCases.kt
@@ -1,0 +1,6 @@
+package domain.validator
+
+data class ValidateUseCases(
+    val validateNumberOfTry: ValidateNumberOfTry = ValidateNumberOfTry(),
+    val validateSuccessiveTokenizer: ValidateSuccessiveTokenizer = ValidateSuccessiveTokenizer()
+)
